Move new test to no_race file
