+++ rest/blip_api_test.go
@@ -3852,63 +3852,3 @@ func TestMultipleOutstandingChangesSubscriptions(t *testing.T) {
 	// activeOneShotStat, activeOneShotStatIsExpected := base.WaitForStat(pullStats.NumPullReplActiveOneShot.Value, 0)
 	// assert.True(t, activeOneShotStatIsExpected, "NumPullReplActiveOneShot=%d instead of expected value of 0", activeOneShotStat)
 }
-
-// TestBlipPusherUpdateDatabase starts a push replication and updates the database underneath the replication.
-// Expect to see the connection closed with an error, instead of continuously panicking.
-func TestBlipPusherUpdateDatabase(t *testing.T) {
-
-	defer base.SetUpTestLogging(base.LevelDebug, base.KeyHTTP, base.KeyHTTPResp, base.KeySync)()
-
-	rtConfig := RestTesterConfig{
-		DatabaseConfig: &DatabaseConfig{},
-		guestEnabled:   true,
-	}
-	rt := NewRestTester(t, &rtConfig)
-	defer rt.Close()
-
-	client, err := NewBlipTesterClientOptsWithRT(t, rt, nil)
-	require.NoError(t, err)
-	defer client.Close()
-
-	var lastPushRevErr atomic.Value
-
-	// Wait for the background updates to finish at the end of the test
-	shouldCreateDocs := base.NewAtomicBool(true)
-	wg := sync.WaitGroup{}
-	wg.Add(1)
-	defer func() {
-		shouldCreateDocs.Set(false)
-		wg.Wait()
-	}()
-
-	// Start the test client creating and pushing documents in the background
-	go func() {
-		for i := 0; shouldCreateDocs.IsTrue(); i++ {
-			// this will begin to error when the database is reloaded underneath the replication
-			_, err := client.PushRev(fmt.Sprintf("doc%d", i), "", []byte(fmt.Sprintf(`{"i":%d}`, i)))
-			if err != nil {
-				lastPushRevErr.Store(err)
-			}
-		}
-		_ = rt.WaitForPendingChanges()
-		wg.Done()
-	}()
-
-	// and wait for a few to be done before we proceed with updating database config underneath replication
-	_, err = rt.WaitForChanges(5, "/db/_changes", "", true)
-	require.NoError(t, err)
-
-	// just change the sync function to cause the database to reload
-	resp, err := rt.UpsertDbConfig("db", DbConfig{Sync: base.StringPtr(`function(doc){console.log("update");}`)})
-	require.NoError(t, err)
-	assertStatus(t, resp, http.StatusCreated)
-
-	// Did we tell the client to close the connection (via HTTP/503)?
-	// The BlipTesterClient doesn't implement reconnect - but CBL resets the replication connection.
-	lastErr, ok := lastPushRevErr.Load().(error)
-	require.True(t, ok)
-	require.Error(t, lastErr)
-	lastErrMsg := lastErr.Error()
-	assert.Contains(t, lastErrMsg, "HTTP 503")
-	assert.Contains(t, lastErrMsg, "Sync Gateway database went away - asking client to reconnect")
-}
